Move Menu object to new namespace
