+++ app/Domain/Utils/MenuItem.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Domain\Menu;
+namespace App\Domain\Utils;
 
 final class MenuItem
 {
